backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.4
Flask-JWT-Extended==4.4.4
Flask-Migrate==4.0.4
Flask-CORS==4.0.0
python-dotenv==1.0.0
reportlab==4.0.7
qrcode==7.4.2
Pillow==10.1.0
anthropic>=0.18.0
requests>=2.31.0
celery[redis]>=5.3.0
redis>=5.0.0
dnspython>=2.4.0
